created logic

+++ src/js/handlers.js
@@ -2,23 +2,25 @@ import onChange from 'on-change';
 import renderItems from './render-items';
 import state from './state';
 
-export default () => {
-  const watchedState = onChange(state, () => {
-    fetch('https://reqres.in/api/unknown?per_page=12')
-      .then((response) => response.json())
-      .then((result) => renderItems(result.data, state));
-  });
+const checkboxes = document.querySelectorAll('input[type="checkbox"]');
+const resetButton = document.querySelector('button[type="reset"]');
+const watchedState = onChange(state, () => {
+  fetch('https://reqres.in/api/unknown?per_page=12')
+    .then((response) => response.json())
+    .then((result) => renderItems(result.data, state));
+});
 
-  const resetButton = document.querySelector('button[type="reset"]');
+export default () => {
   resetButton.addEventListener('click', (event) => {
     event.preventDefault();
-    localStorage.clear();
-    Object.keys(state).forEach((key) => {
-      watchedState[key] = true;
-    });
+    if (Object.values(state).includes(false)) {
+      localStorage.clear();
+      Object.keys(state).forEach((key) => {
+        watchedState[key] = true;
+      });
+    }
   });
 
-  const checkboxes = document.querySelectorAll('input[type="checkbox"]');
   checkboxes.forEach((input) => {
     input.addEventListener('click', (event) => {
       const data = event.target.dataset.checkbox;
